+++ app/src/main/java/com/android/people/quotesandroidapp/provider/status/StatusModel.java
@@ -1,11 +1,8 @@
 package com.android.people.quotesandroidapp.provider.status;
 
-import com.android.people.quotesandroidapp.provider.base.BaseModel;
-
-import java.util.Date;
-
-import android.support.annotation.NonNull;
 import android.support.annotation.Nullable;
+
+import com.android.people.quotesandroidapp.provider.base.BaseModel;
 
 /**
  * Status for every quote 
@@ -15,7 +12,7 @@ public interface StatusModel extends BaseModel {
     /**
      * Id for the quote
      */
-    int getQuoteid();
+    long getQuoteid();
 
     /**
      * Is the quote favorited?
